requirements.txt
fastapi==0.111.1
uvicorn==0.25.0
pandas==2.1.1
numpy==1.26.2
python-dateutil==2.9.2
jinja2==3.1.2
python-multipart==0.0.6
